Add files via upload
this contain comments and the working makefile. if you call make in the project directory it will make and move everything as needed i just had to do it recursively. the clean works from the outer make file as well as long as out call make from the outer Makefile. if you do a once over of the code it should be ready for submission.