Add files via upload
i have added the button functionality to the blocks/characters. when you are in/next to a block a prompt will go up on the screen telling the user what button to do what action (ex: click 'u' to pick up a blue block). if you go over the block and don't pick it up the block char/block will be re-drawn to the screen.